+++ package.json
@@ -12,6 +12,7 @@
     "formik": "^2.2.9",
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
+    "react-bootstrap": "1.6.4",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
     "react-router-dom": "5.3.0",
