+++ src/App.css
@@ -14,6 +14,15 @@ label {
   margin: auto;
 }
 
+.add-reminder-modal-btn-text {
+  font-size: 1.4rem;
+  padding: 0;
+  height: 2.4rem;
+  width: 2.4rem;
+  line-height: 2rem;
+  display: inline-block;
+}
+
 .card {
   background-color: var(--light-gray);
   padding: 1.2rem 1.6rem 1.9rem;
@@ -62,4 +71,13 @@ label {
   to {
     text-shadow: 0 0 20px #fff, 0 0 30px #8ee600, 0 0 40px #8ee600, 0 0 50px #8ee600, 0 0 60px #8ee600, 0 0 70px #8ee600, 0 0 80px #8ee600;
   }
+}
+
+.new-reminder-form {
+  padding: 0;
+  margin-top: 0;
+  margin-bottom: 0;
+  border: none;
+  box-shadow: none;
+  background-color: transparent;
 }
